requirements.txt
requests
gradio
python-dotenv